+++ systemtest-utils/src/main/java/com.bitmovin.analytics.systemtest.utils/EventDataUtils.kt
@@ -3,6 +3,7 @@ package com.bitmovin.analytics.systemtest.utils
 import com.bitmovin.analytics.data.EventData
 
 object EventDataUtils {
+    // TODO: refactor to return new list, instead of modifying the input list
     fun filterNonDeterministicEvents(eventDataList: MutableList<EventData>) {
         // We filter for qualitychange and buffering events
         // since they are non deterministic and would probably make the test flaky